TCP/UDP unit tests added.
