+++ src/al.c
@@ -614,7 +614,7 @@ ssize_t al_sendto(int32_t sockfd, const void * buf, size_t len, const struct soc
     fret = sendto(sockfd, buf, len, flags, peer_sockaddr, sizeof(peer_sockaddr));
     if(-1 == fret)
     {
-        db_perror("al_sendto");
+        db_perror("al_sendto"); 
     }
 
     return fret;
